directInput
changed the input handler from windows functions to directinput - again
still adding in stuff from rastertek.com

the only original thing i did was using the mouse movement to rotate the
camera. NOTE: the code should be moved to the frame method of the camera
class. ALSO: the GetDirection method of the camera doesn't work
properly, should tweak with that a bit.